+++ include/LIEF/PE/EnumToString.hpp
@@ -30,7 +30,6 @@ enum class ACCELERATOR_FLAGS : size_t;
 enum class ACCELERATOR_VK_CODES : size_t;
 enum class ALGORITHMS : uint32_t;
 enum class CODE_PAGES : size_t;
-enum class DATA_DIRECTORY : size_t;
 enum class DIALOG_BOX_STYLES : size_t;
 enum class EXTENDED_WINDOW_STYLES : size_t;
 enum class FIXED_VERSION_FILE_FLAGS : size_t;
@@ -56,7 +55,6 @@ enum class WINDOW_STYLES : size_t;
 enum class WIN_VERSION : size_t;
 
 LIEF_API const char* to_string(PE_TYPE e);
-LIEF_API const char* to_string(DATA_DIRECTORY e);
 
 LIEF_API const char* to_string(SECTION_CHARACTERISTICS e);
 LIEF_API const char* to_string(PE_SECTION_TYPES e);