+++ include/LIEF/PE/enums.hpp
@@ -184,27 +184,6 @@ enum class WeakExternalCharacteristics: size_t  {
 };
 
 
-enum class DATA_DIRECTORY: size_t  {
-  EXPORT_TABLE            = 0,
-  IMPORT_TABLE            = 1,
-  RESOURCE_TABLE          = 2,
-  EXCEPTION_TABLE         = 3,
-  CERTIFICATE_TABLE       = 4,
-  BASE_RELOCATION_TABLE   = 5,
-  DEBUG                   = 6,
-  ARCHITECTURE            = 7,
-  GLOBAL_PTR              = 8,
-  TLS_TABLE               = 9,
-  LOAD_CONFIG_TABLE       = 10,
-  BOUND_IMPORT            = 11,
-  IAT                     = 12,
-  DELAY_IMPORT_DESCRIPTOR = 13,
-  CLR_RUNTIME_HEADER      = 14,
-  RESERVED                = 15,
-
-  NUM_DATA_DIRECTORIES    = 16
-};
-
 enum class ImportType: size_t  {
   IMPORT_CODE  = 0,
   IMPORT_DATA  = 1,